+++ native_library/android/FreKotlinExampleANE/FreKotlin/src/main/java/com/tuarua/frekotlin/Extensions.kt
@@ -18,13 +18,12 @@
 
 package com.tuarua.frekotlin
 
+import android.graphics.Color
 import com.adobe.fre.FREArray
 import com.adobe.fre.FREContext
 import com.adobe.fre.FREObject
-import com.tuarua.frekotlin.geom.FrePointKotlin
-import com.tuarua.frekotlin.geom.Point
 import java.util.*
-import kotlin.collections.ArrayList
+//import kotlin.collections.ArrayList
 
 typealias FREArgv = ArrayList<FREObject>
 fun FREContext.sendEvent(name: String, value: String) {
@@ -45,226 +44,336 @@ inline fun <T> T.guard(block: T.() -> Unit): T {
 }
 
 fun Double(freObject: FREObject?): Double? {
-    val v = FreObjectKotlin(freObject = freObject).value
-    return (v as? Int)?.toDouble() ?: v as Double?
-}
-
-fun Double(freObjectKotlin: FreObjectKotlin?): Double? {
-    val v = freObjectKotlin?.value
+    val v = FreKotlinHelper.getAsObject(freObject)
     return (v as? Int)?.toDouble() ?: v as Double?
 }
 
 fun Float(freObject: FREObject?): Float? {
-    val v = FreObjectKotlin(freObject = freObject).value
+    val v = FreKotlinHelper.getAsObject(freObject)
     return (v as? Int)?.toFloat() ?: (v as Double?)?.toFloat()
 }
 
-fun Float(freObjectKotlin: FreObjectKotlin?): Float? {
-    val v = freObjectKotlin?.value
-    return (v as? Int)?.toFloat() ?: (v as Double?)?.toFloat()
-}
-
-fun IntArray(freArray: FREArray?): IntArray? {
-    if (freArray != null) {
-        val al = FreArrayKotlin(freArray).value
-        val count = al.count()
-        val kotArr = IntArray(count)
-        for (i in 0 until count) {
-            val v = al[i]
-            if (v is Int) {
-                kotArr.set(index = i,value = v)
-            }else{
-                return null
-            }
-        }
-        return kotArr
-    }
-    return null
-}
-
-fun IntArray(freObject: FREObject?): IntArray? {
-    if (freObject != null) {
-        val al = FreArrayKotlin(freObject).value
-        val count = al.count()
-        val kotArr = IntArray(count)
-        for (i in 0 until count) {
-            val v = al[i]
-            if (v is Int) {
-                kotArr.set(index = i,value = v)
-            }else{
-                return null
-            }
-        }
-        return kotArr
-    }
-    return null
-}
-
-fun IntArray(freArrayKotlin: FreArrayKotlin?): IntArray? {
-    if (freArrayKotlin != null) {
-        val al = freArrayKotlin.value
-        val count = al.count()
-        val kotArr = IntArray(count)
-        for (i in 0 until count) {
-            val v = al[i]
-            if (v is Int) {
-                kotArr.set(index = i,value = v)
-            }else{
-                return null
-            }
-        }
-        return kotArr
-    }
-    return null
-}
-
-fun DoubleArray(freArray: FREArray?): DoubleArray? {
-    if (freArray != null) {
-        val al = FreArrayKotlin(freArray).value
-        val count = al.count()
-        val kotArr: DoubleArray = kotlin.DoubleArray(count)
-        for (i in 0 until count) if (al[i] is Double) {
-            val v: Double = al[i] as Double
-            kotArr[i] = v
-        } else {
-            return null
-        }
-        return kotArr
-    }
-    return null
-}
-
-fun DoubleArray(freObject: FREObject?): DoubleArray? {
-    if (freObject != null) {
-        val al = FreArrayKotlin(freObject).value
-        val count = al.count()
-        val kotArr: DoubleArray = kotlin.DoubleArray(count)
-        for (i in 0 until count) if (al[i] is Double) {
-            val v: Double = al[i] as Double
-            kotArr[i] = v
-        } else {
-            return null
-        }
-        return kotArr
-    }
-    return null
-}
-
-fun DoubleArray(freArrayKotlin: FreArrayKotlin?): DoubleArray? {
-    if (freArrayKotlin != null) {
-        val al = freArrayKotlin.value
-        val count = al.count()
-        val kotArr: DoubleArray = kotlin.DoubleArray(count)
-        for (i in 0 until count) if (al[i] is Double) {
-            val v: Double = al[i] as Double
-            kotArr[i] = v
-        } else {
-            return null
-        }
-        return kotArr
-    }
-    return null
-}
-
+@Throws(FreException::class)
 fun BooleanArray(freArray: FREArray?): BooleanArray? {
     if (freArray != null) {
-        val al = FreArrayKotlin(freArray).value
-        val count = al.count()
-        val kotArr = BooleanArray(count)
-        for (i in 0 until count) if (al[i] is Boolean) {
-            val v: Boolean = al[i] as Boolean
-            kotArr[i] = v
-        } else {
-            return null
+        val count = freArray.length.toInt()
+        val kotArr: BooleanArray = kotlin.BooleanArray(count)
+        for (i in 0 until count) {
+            val fre = freArray.getObjectAt(i.toLong())
+            val b = Boolean(FreObjectKotlin(fre).rawValue)
+            when {
+                b != null -> kotArr[i] = b
+                else -> return null
+            }
         }
         return kotArr
     }
     return null
 }
 
+@Throws(FreException::class)
 fun BooleanArray(freObject: FREObject?): BooleanArray? {
     if (freObject != null) {
-        val al = FreArrayKotlin(freObject).value
-        val count = al.count()
-        val kotArr = BooleanArray(count)
-        for (i in 0 until count) if (al[i] is Boolean) {
-            val v: Boolean = al[i] as Boolean
-            kotArr[i] = v
-        } else {
-            return null
+        return BooleanArray(FREArray(freObject))
+    }
+    return null
+}
+
+@Throws(FreException::class)
+fun DoubleArray(freArray: FREArray?): DoubleArray? {
+    if (freArray != null) {
+        val count = freArray.length.toInt()
+        val kotArr: DoubleArray = kotlin.DoubleArray(count)
+        for (i in 0 until count) {
+            val fre = freArray.getObjectAt(i.toLong())
+            val b = Double(FreObjectKotlin(fre).rawValue)
+            when {
+                b != null -> kotArr[i] = b
+                else -> return null
+            }
         }
         return kotArr
     }
     return null
 }
 
-fun BooleanArray(freArrayKotlin: FreArrayKotlin?): BooleanArray? {
-    if (freArrayKotlin != null) {
-        val al = freArrayKotlin.value
-        val count = al.count()
-        val kotArr = BooleanArray(count)
-        for (i in 0 until count) if (al[i] is Boolean) {
-            val v: Boolean = al[i] as Boolean
-            kotArr[i] = v
-        } else {
-            return null
+@Throws(FreException::class)
+fun DoubleArray(freObject: FREObject?): DoubleArray? {
+    if (freObject != null) {
+        return DoubleArray(FREArray(freObject))
+    }
+    return null
+}
+
+@Throws(FreException::class)
+fun IntArray(freArray: FREArray?): IntArray? {
+    if (freArray != null) {
+        val count = freArray.length.toInt()
+        val kotArr: IntArray = kotlin.IntArray(count)
+        for (i in 0 until count) {
+            val fre = freArray.getObjectAt(i.toLong())
+            val b = Int(FreObjectKotlin(fre).rawValue)
+            when {
+                b != null -> kotArr[i] = b
+                else -> return null
+            }
         }
         return kotArr
     }
     return null
 }
+
 @Throws(FreException::class)
-fun Int.toFREObject():FREObject? {
-    return try {
-        FreObjectKotlin(this).rawValue
-    } catch (e: FreException) {
-        e.getError(Thread.currentThread().stackTrace)
+fun IntArray(freObject: FREObject?): IntArray? {
+    if (freObject != null) {
+        return IntArray(FREArray(freObject))
+    }
+    return null
+}
+
+@Suppress("UNCHECKED_CAST", "unused")
+fun <String> List(freArray: FREArray?): List<String> {
+    if (freArray != null) {
+        return freArray.let { FreKotlinHelper.getAsObject(it) as List<String> }
+    }
+    return listOf()
+}
+
+//private var deviceList: List<String>? = null
+
+@Suppress("UNCHECKED_CAST", "unused")
+fun <String, Any>Map(freObject: FREObject?): Map<String, Any>? {
+    if (freObject != null) {
+        return FreKotlinHelper.getAsObject(freObject) as Map<String, Any>?
+    }
+    return null
+}
+
+fun Int(freObject: FREObject?): Int? = FreKotlinHelper.getAsObject(freObject) as Int?
+
+fun String(freObject: FREObject?): String? = FreKotlinHelper.getAsObject(freObject).toString()
+
+fun Boolean(freObject: FREObject?): Boolean? = FreKotlinHelper.getAsObject(freObject) as Boolean?
+
+fun Date(freObject: FREObject?): Date? = FreKotlinHelper.getAsObject(freObject) as Date?
+
+@Throws(FreException::class)
+fun FREObject?.call(method: String, vararg args: Any): FREObject? {
+    val rv = this ?: return null
+    return FreKotlinHelper.call(rv, method, args)
+}
+
+@Throws(FreException::class)
+fun FREObject?.call(method: String): FREObject? {
+    val rv = this ?: return null
+    return FreKotlinHelper.call(rv, method)
+}
+
+@Throws(FreException::class)
+fun FREObject?.call(method: String, vararg args: FREObject): FREObject? {
+    val rv = this ?: return null
+    return FreKotlinHelper.call(rv, method, args)
+}
+
+var FREObject.type: FreObjectTypeKotlin
+    get() {
+        return FreKotlinHelper.getType(this)
+    }
+    set(value) {}
+
+@Throws(FreException::class)
+fun FREObject(name: String, vararg args: Any): FREObject {
+    val argsArr = arrayOfNulls<FREObject>(args.size)
+    for (i in args.indices) {
+        argsArr[i] = FreObjectKotlin(args[i]).rawValue
+    }
+    try {
+        return FREObject.newObject(name, argsArr)
     } catch (e: Exception) {
-        FreException(e).getError(Thread.currentThread().stackTrace)
+        throw FreException(e, "cannot create new object named $name")
+    }
+}
+
+fun FREArray(freObject: FREObject): FREArray {
+    return freObject as FREArray
+}
+
+@Throws(FreException::class)
+fun FREArray(value: IntArray): FREArray {
+    val rv = FREArray.newArray(value.size)
+    for (i in value.indices) {
+        rv.set(i, value[i])
+    }
+    return rv
+}
+
+@Throws(FreException::class)
+fun FREArray(value: DoubleArray): FREArray {
+    val rv = FREArray.newArray(value.size)
+    for (i in value.indices) {
+        rv.set(i, value[i])
+    }
+    return rv
+}
+
+fun FREArray(value: BooleanArray): FREArray {
+    val rv = FREArray.newArray(value.size)
+    for (i in value.indices) {
+        rv.set(i, value[i])
+    }
+    return rv
+}
+
+@Throws(FreException::class)
+fun FREArray(value: List<Any>): FREArray {
+    val rv = FREArray.newArray(value.size)
+    for (i in value.indices) {
+        rv.set(i, value[i])
+    }
+    return rv
+}
+
+@Throws(FreException::class)
+fun FREArray.at(index: Int): FREObject? {
+    return this.getObjectAt(index.toLong())
+}
+
+@Throws(FreException::class)
+fun FREArray.set(index: Int, value: Any?) {
+    this.setObjectAt(index.toLong(), FreObjectKotlin(value).rawValue)
+}
+
+@Throws(FreException::class)
+fun FREObject.setProp(name: String, value: Any?) {
+    if (value is FREObject) {
+        FreKotlinHelper.setProperty(this, name, value)
+        return
+    }
+
+    if (value is FreObjectKotlin) {
+        FreKotlinHelper.setProperty(this, name, value.rawValue)
+        return
+    }
+
+    if (value is String) {
+        FreKotlinHelper.setProperty(this, name, value.toFREObject())
+        return
+    }
+    if (value is Int) {
+        FreKotlinHelper.setProperty(this, name, value.toFREObject())
+        return
+    }
+
+    if (value is Double) {
+        FreKotlinHelper.setProperty(this, name, value.toFREObject())
+        return
+    }
+
+    if (value is Long) {
+        FreKotlinHelper.setProperty(this, name, value.toFREObject())
+        return
+    }
+
+    if (value is Short) {
+        FreKotlinHelper.setProperty(this, name, value.toFREObject())
+        return
+    }
+
+    if (value is Boolean) {
+        FreKotlinHelper.setProperty(this, name, value.toFREObject())
+        return
+    }
+
+    if (value is Date) {
+        FreKotlinHelper.setProperty(this, name, value.toFREObject())
+        return
+    }
+
+    if (value is Any) {
+        //Log.e(TAG, "any is an Any - NOT FOUND")
+        return
     }
 }
 
 @Throws(FreException::class)
-fun Boolean.toFREObject():FREObject? {
-    return try {
-        FreObjectKotlin(this).rawValue
-    } catch (e: FreException) {
-        e.getError(Thread.currentThread().stackTrace)
+fun FREObject.getProp(name: String): FREObject? {
+    return FreKotlinHelper.getProperty(this, name)
+}
+
+fun FREObject.toColor(alpha: Int = 255): Int {
+    if (this.type != FreObjectTypeKotlin.INT) return 0
+    val freColor = Int(this) ?: return 0
+    return Color.argb(alpha, Color.red(freColor), Color.green(freColor), Color.blue(freColor))
+}
+
+fun FREObject.toHSV(alpha: Int = 255): Float {
+    if (this.type != FreObjectTypeKotlin.INT) return 0.0F
+    val hsv = FloatArray(3)
+    Color.colorToHSV(toColor(alpha), hsv)
+    return hsv[0]
+}
+
+// ************ toFREObject **************//
+@Throws(FreException::class)
+fun Int.toFREObject(): FREObject? {
+    try {
+        return FREObject.newObject(this)
     } catch (e: Exception) {
-        FreException(e).getError(Thread.currentThread().stackTrace)
+        throw FreException(e, "cannot create new object from Int")
     }
 }
 
 @Throws(FreException::class)
-fun String.toFREObject():FREObject? {
-    return try {
-        FreObjectKotlin(this).rawValue
-    } catch (e: FreException) {
-        e.getError(Thread.currentThread().stackTrace)
+fun Short.toFREObject(): FREObject? {
+    try {
+        return FREObject.newObject(this.toInt())
     } catch (e: Exception) {
-        FreException(e).getError(Thread.currentThread().stackTrace)
+        throw FreException(e, "cannot create new object from Int")
     }
 }
 
 @Throws(FreException::class)
-fun Double.toFREObject():FREObject? {
-    return try {
-        FreObjectKotlin(this).rawValue
-    } catch (e: FreException) {
-        e.getError(Thread.currentThread().stackTrace)
+fun Boolean.toFREObject(): FREObject? {
+    try {
+        return FREObject.newObject(this)
     } catch (e: Exception) {
-        FreException(e).getError(Thread.currentThread().stackTrace)
+        throw FreException(e, "cannot create new object from Boolean")
     }
 }
 
 @Throws(FreException::class)
-fun Float.toFREObject():FREObject? {
-    return try {
-        FreObjectKotlin(this).rawValue
-    } catch (e: FreException) {
-        e.getError(Thread.currentThread().stackTrace)
+fun String.toFREObject(): FREObject? {
+    try {
+        return FREObject.newObject(this)
     } catch (e: Exception) {
-        FreException(e).getError(Thread.currentThread().stackTrace)
+        throw FreException(e, "cannot create new object from String")
+    }
+}
+
+@Throws(FreException::class)
+fun Double.toFREObject(): FREObject? {
+    try {
+        return FREObject.newObject(this)
+    } catch (e: Exception) {
+        throw FreException(e, "cannot create new object from Double")
+    }
+}
+
+@Throws(FreException::class)
+fun Long.toFREObject(): FREObject? {
+    try {
+        return FREObject.newObject(this.toDouble())
+    } catch (e: Exception) {
+        throw FreException(e, "cannot create new object from Double")
+    }
+}
+
+@Throws(FreException::class)
+fun Float.toFREObject(): FREObject? {
+    try {
+        return FREObject.newObject(this.toDouble())
+    } catch (e: Exception) {
+        throw FreException(e, "cannot create new object from Double")
     }
 }
 
@@ -278,16 +387,3 @@ fun Date.toFREObject(): FREObject? {
         FreException(e).getError(Thread.currentThread().stackTrace)
     }
 }
-
-fun IntArray.toFREObject():FREArray? {
-    return FreArrayKotlin(this).rawValue
-}
-
-fun Int(freObject: FREObject?): Int? = FreObjectKotlin(freObject = freObject).value as Int?
-fun Int(freObjectKotlin: FreObjectKotlin?): Int? = freObjectKotlin?.value as Int?
-fun String(freObject: FREObject?): String? = FreObjectKotlin(freObject = freObject).value?.toString()
-fun String(freObjectKotlin: FreObjectKotlin?): String? = freObjectKotlin?.value?.toString()
-fun Boolean(freObject: FREObject?): Boolean? = FreObjectKotlin(freObject = freObject).value as Boolean?
-fun Boolean(freObjectKotlin: FreObjectKotlin?): Boolean? = freObjectKotlin?.value as Boolean?
-fun Date(freObject: FREObject?): Date? = FreObjectKotlin(freObject = freObject).value as Date?
-fun Date(freObjectKotlin: FreObjectKotlin?): Date? = freObjectKotlin?.value as Date?