+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+### 0.0.10 
+- Extensions for conversion to ArrayList, Map
+- Extensions for FREObject, FREArray
+
 ### 0.0.9  
 - Extensions for conversion to IntArray, DoubleArray, BooleanArray